renamed db component to list
